# lib/api/openapi.yaml
openapi: 3.0.3
info:
  title: Backend PFE API
  version: 1.0.0
paths:
...
  /api/v1/accounts/login/:
    post:
      operationId: v1_accounts_login_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomTokenObtainPair'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CustomTokenObtainPair'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CustomTokenObtainPair'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomTokenObtainPair'
          description: ''
...
  /api/v1/requests/:
    get:
      operationId: v1_requests_list
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceRequestList'
          description: ''
    post:
      operationId: v1_requests_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequestCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ServiceRequestCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ServiceRequestCreate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequestCreate'
          description: ''
...
components:
  schemas:
    CustomTokenObtainPair:
      type: object
      properties:
        phone:
          type: string
          writeOnly: true
        password:
          type: string
          writeOnly: true
      required:
      - password
      - phone
    DispatchStateEnum:
      enum:
      - pending
      - offered
      - accepted
      - expired
      - cancelled
      type: string
      description: |-
        * `pending` - Pending
        * `offered` - Offered
        * `accepted` - Accepted
        * `expired` - Expired
        * `cancelled` - Cancelled
...
    ServiceRequestCreate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        vehicle:
          type: integer
        vehicle_name:
          type: string
          readOnly: true
        service_type:
          type: string
          maxLength: 100
        description:
          type: string
        image:
          type: string
          format: uri
          nullable: true
        image_url:
          type: string
          readOnly: true
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        address:
          type: string
          maxLength: 255
        status:
          allOf:
          - $ref: '#/components/schemas/Status416Enum'
          readOnly: true
        dispatch_state:
          allOf:
          - $ref: '#/components/schemas/DispatchStateEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - dispatch_state
      - id
      - image_url
      - latitude
      - longitude
      - service_type
      - status
      - vehicle
      - vehicle_name
...
    ServiceRequestList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        vehicle:
          type: integer
          nullable: true
        vehicle_name:
          type: string
          readOnly: true
        service_type:
          type: string
          maxLength: 100
        description:
          type: string
        status:
          $ref: '#/components/schemas/Status416Enum'
        dispatch_state:
          $ref: '#/components/schemas/DispatchStateEnum'
        customer_first_name:
          type: string
          readOnly: true
        customer_last_name:
          type: string
          readOnly: true
        customer_phone:
          type: string
          readOnly: true
        provider_first_name:
          type: string
          readOnly: true
        provider_last_name:
          type: string
          readOnly: true
        image:
          type: string
          format: uri
          nullable: true
        image_url:
          type: string
          readOnly: true
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        address:
          type: string
          maxLength: 255
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - customer_first_name
      - customer_last_name
      - customer_phone
      - id
      - image_url
      - latitude
      - longitude
      - provider_first_name
      - provider_last_name
      - service_type
      - vehicle_name
    Status416Enum:
      enum:
      - pending
      - accepted
      - provider_on_the_way
      - arrived
      - completed
      - cancelled
      type: string
      description: |-
        * `pending` - Pending
        * `accepted` - Accepted
        * `provider_on_the_way` - Provider On The Way
        * `arrived` - Arrived
        * `completed` - Completed
        * `cancelled` - Cancelled
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT